feat(log): write logs to file when polling for job status
